Creating modular endpoints.
